continuation refactor db _ still no check because of IT issues
